+++ cdcrush/prog/TaskRestoreTrack.cs
@@ -18,6 +18,10 @@ class TaskRestoreTrack : lib.task.CTask
 	RestoreParams p;
 	CueTrack track;
 
+	
+	bool isOgg = false;
+	bool isFlac = false;
+
 	string crushedTrackPath;  // Autocalculated
 
 	// --
@@ -49,6 +53,10 @@ class TaskRestoreTrack : lib.task.CTask
 			ecm.onComplete = (s) => {
 				if(s){
 					deleteOldFile();
+					if(!checkTrackMD5()) {
+						fail(msg:"MD5 checksum is wrong!");
+						return;
+					}
 					complete();
 				}else{
 					fail(msg:ecm.ERROR);
@@ -59,16 +67,31 @@ class TaskRestoreTrack : lib.task.CTask
 		}
 		else
 		{
+			if(Path.GetExtension(track.storedFileName) == ".ogg") {
+				isOgg = true;
+			}else{
+				isFlac = true; // must be flac
+			}
+
 			var ffmp = new FFmpeg(CDCRUSH.FFMPEG_PATH);
 			ffmp.onComplete = (s) => {
 				if(s){
-					deleteOldFile();
-					correctPCMSize();
+					deleteOldFile(); // Don't need it
+					if(isOgg) {
+						correctPCMSize();
+					}
+					if(isFlac) {
+						if(!checkTrackMD5()){
+							fail(msg:"MD5 checksum is wrong!");
+							return;
+						}
+					}
 					complete();
 				}else{
 					fail(msg:ffmp.ERROR);
 				}
 			};
+
 			ffmp.audioToPCM(crushedTrackPath);
 			killExtra = () => ffmp.kill();
 		}
@@ -78,20 +101,47 @@ class TaskRestoreTrack : lib.task.CTask
 	// --
 	void deleteOldFile()
 	{
+		if(CDCRUSH.FLAG_KEEP_TEMP) return;
 		File.Delete(crushedTrackPath);
 	}// -----------------------------------------
 
-	// Implies restored OK and file exists/
+	
+	// -
+	// Fix the filesize of the restored track
+	// This is only when restoring from .OGG files, .FLAC seems to be fine by default.
 	void correctPCMSize()
 	{
-		//log("SIZE BEFORE " + new FileInfo(track.workingFile).Length); -> OK WORKS
 		using(FileStream fileStream = new FileStream(track.workingFile,FileMode.Open, FileAccess.Write))
 		{
 			fileStream.SetLength(track.byteSize);
 		}
-		//log("SIZE AFTER " + new FileInfo(track.workingFile).Length); -> OK WORKS
+	}// -----------------------------------------
+
+	// Check the track's generated MD5 against the original MD5
+	// It can take some time if it's a big file.
+	bool checkTrackMD5()
+	{
+		// Only check if requested? TODO.
+		return true; 
+
+		if(track.md5==null){
+			LOG.log("Cannot check MD5 against original, this is an old version archive where md5 was not calculated.");
+			return true;
+		}
+
+		string newMD5 = "";
+
+		// Before compressing the tracks, get and store the MD5 of the track
+		using(var md5 = System.Security.Cryptography.MD5.Create())
+		{
+			using(var str = File.OpenRead(track.workingFile))
+			{
+				newMD5 = System.BitConverter.ToString(md5.ComputeHash(str)).Replace("-","").ToLower();
+			}
+		}
+
+		return (track.md5==newMD5);
 	}// -----------------------------------------
 
 }// -- end class
-	
 }// -- end namespace
